+++ wire/lib/src/nix_log.rs
@@ -57,13 +57,13 @@ impl Trace for Internal {
                 };
 
                 match level {
-                    Level::Info | Level::Talkative | Level::Chatty => {
-                        event!(tracing_level::INFO, "{text}");
-                    }
+                    Level::Info => event!(tracing_level::INFO, "{text}"),
                     Level::Warn | Level::Notice => event!(tracing_level::WARN, "{text}"),
                     Level::Error => event!(tracing_level::ERROR, "{text}"),
                     Level::Debug => event!(tracing_level::DEBUG, "{text}"),
-                    Level::Vomit => event!(tracing_level::TRACE, "{text}"),
+                    Level::Vomit | Level::Talkative | Level::Chatty => {
+                        event!(tracing_level::TRACE, "{text}");
+                    }
                 }
             }
             Action::Stop => {}